Model/old_model/requirements.txt
tensorflow==2.7.0
tensorflow-gpu==2.7.0
keras==2.7.0
matplotlib==3.5.0
librosa==0.8.1
numpy==1.20.3
Pillow==8.4.0